Added test coverage
