+++ src/test/resources/algorithms/testing/99.99/glossary/554540cfe4b0426fced42655.json
@@ -1,5 +1,5 @@
 {
-  "name": "carcinoma",
+  "name": "Carcinoma",
   "definition": "Placeholder for definition",
   "last_modified": "2015-07-23T17:36:43.477Z"
 }
