+++ src/test/resources/algorithms/testing/99.99/glossary/terms.txt
@@ -1,19 +1,13 @@
 Adjacent structures~55a80262e4b05cd0cddb99b8
-Collaborative stage data collection system~554540cfe4b0426fced42655
-Connective tissue~55a800abe4b05cd0cddb978e
+Adjacent tissue(s), NOS~55a800abe4b05cd0cddb978e
+Carcinoma~554540cfe4b0426fced42655
 Cortex~546cf5aee4b0d965832a94bf
-Cortical~546cf5aee4b0d965832a94bf
-ITC~55dcb26c1ef5576bf30ccc10
-ITCs~55dcb26c1ef5576bf30ccc10
 Isolated tumor cells~55dcb26c1ef5576bf30ccc10
-Level VA~55a9a501e4b05cd0cddc710a
-Level VB~55a9a501e4b05cd0cddc710a
+Level V lymph nodes~55a9a501e4b05cd0cddc710a
 Masticator space~55ba2b9b1ef5572aac7b2910
 Medulla~55b12ee91ef5572aac76e67a
-Medullary~55b12ee91ef5572aac76e67a
 Parenchyma~546f598de4b0d965832bb0c4
 Postcricoid region~5505c8e7e4b0c48f31d70838
-Satellite nodules~55dcb1521ef5576bf30cca55
 Stroma~55097bbee4b0c48f31d8967a
 Summary Stage~550adaede4b0c48f31d94520
 Tumor deposits~55dcb1521ef5576bf30cca55
