+++ src/test/resources/algorithms/testing/99.99/tables/ids.txt
@@ -1,0 +1,62 @@
+ajcc6_exclusions_ppd
+ajcc6_m_codes
+ajcc6_n_codes
+ajcc6_stage_codes
+ajcc6_stage_qaw
+ajcc6_t_codes
+ajcc6_year_validation
+ajcc7_inclusions_tpb
+ajcc7_m_codes
+ajcc7_n_codes
+ajcc7_stage_codes
+ajcc7_stage_uaz
+ajcc7_t_codes
+ajcc7_year_validation
+ajcc_descriptor_codes
+ajcc_mdescriptor_cleanup
+ajcc_ndescriptor_cleanup
+ajcc_tdescriptor_cleanup
+behavior
+cs_input_version_original
+cs_year_validation
+extension_bdi
+extension_eval_cpa
+grade
+histology
+lvi
+mets_eval_ipa
+mets_hpe
+nodes_dbu
+nodes_eval_epa
+nodes_exam_gpa
+nodes_pos_fpa
+primary_site
+schema_selection_urethra
+size_apa
+ss_codes
+ssf1_jpd
+ssf2_kna
+ssf3_lna
+ssf4_mna
+ssf5_nna
+ssf6_ona
+ssf7_snb
+ssf8_snc
+ssf9_snd
+ssf10_sne
+ssf11_snf
+ssf12_sng
+ssf13_snh
+ssf14_sni
+ssf15_snj
+ssf16_snk
+ssf17_snl
+ssf18_snm
+ssf19_snn
+ssf20_sno
+ssf21_snp
+ssf22_snq
+ssf23_snr
+ssf24_sns
+ssf25_snt
+summary_stage_rpa
